Revert "Redirect to /login for restricted pages"
This reverts commit 48d7b0f22be967e14720a71f7f33825651911f9d.
